+++ docs/changelog.rst
@@ -5,6 +5,10 @@ Changelog
 * :feature:`2384` Users will now see their loopring balances on dashboard nested underneath the Ethereum balances.
 * :feature:`1448` When querying trades, deposits and withdrawals the entries that have already been queried will now be instantly shown to the user, while waiting for the query of the latest entries to complete.
 * :feature:`1799` Modules will now be dynamically activated/deactivated at the moment the user modifies the settings from the frontend. Restarts of the app will no longer be necessary.
+
+* :release:`1.14.2 <2021-02-24>`
+* :bug:`2399` Users will now see a warning if the loopring module is not activated when adding an API key, and balances will be fetched automatically if it is.
+* :bug:`2151` Users will now see the datetime picker properly displaying the selected date when editing ledger actions.
 * :bug:`2405` Legacy bitcoin address balances and xpub derivation should now work properly again after blockchain.info decided to yolo change their api response format.
 * :bug:`2400` Loopring balances should now be queried properly for users who own USDT.
 * :bug:`2398` An edge case of Kucoin historical trade query parsing is fixed. So now even users with some specific ids in their trades will be able to query history properly for Kucoin.